address ms compiler warnings

+++ src/geometry.cpp
@@ -89,8 +89,8 @@ std::unique_ptr<ColorMeshGeometry> Geometry::createSphereColorMeshGeometry(const
         if (isBottom) {
             j = lonOffset;
             mesh.indices.emplace_back(j);
-            mesh.indices.emplace_back(mesh.vertices.size()-2);
-            mesh.indices.emplace_back(mesh.vertices.size()-1);
+            mesh.indices.emplace_back(static_cast<uint32_t>(mesh.vertices.size())-2);
+            mesh.indices.emplace_back(static_cast<uint32_t>(mesh.vertices.size())-1);
             while (j<lonOffset+lonIntervals) {
                 mesh.indices.emplace_back(j);
                 mesh.indices.emplace_back(mesh.vertices.size()-1);
